requirements.txt
aniso8601==7.0.0
asgiref==3.5.2
Django==3.2.9
django-cors-headers==3.13.0
django-countries==7.4.2
django-filter==22.1
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
django-stubs==1.13.0
django-stubs-ext==0.7.0
graphene==2.1.9
graphene-django==2.15.0
graphene-stubs==0.15
graphql-core==2.3.2
graphql-relay==2.0.1
mypy==0.991
mypy-extensions==0.4.3
Pillow==9.3.0
promise==2.3
PyJWT==1.7.1
python-magic==0.4.27
pytz==2022.6
pyuca==1.2
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
tomli==2.0.1
types-pytz==2022.6.0.1
types-PyYAML==6.0.12.2
typing_extensions==4.4.0
